Initial commit.
This is the initial commit of a compiler for the Tiger programming
language as presented in Appel's book 'Modern Compiler Implementation in C'.
The backend is LLVM, the Low Level Virtual Machine <http://www.llvm.org>.
The main objectives are:

- complete and correct support for the whole language.
- support for garbage collection (using LLVM gcroot infrastructure).
- support for exceptions (using LLVM exception handling).

I was inspired to begin this project by Paul Govereau's Tiger
compiler for C-- <http://www.cs.tufts.edu/~nr/c--/>.
This compiler supports the complete Tiger language, plus GC and
exceptions, and targets the C-- intermediate language. Unfortunately,
C-- seems to have died, and Govereau's compiler with it. We hope that
once this project is complete it will serve as an example of how to
do some things in LLVM, such as GC and exceptions for which it seems
there are no examples easily available.

The current status is PRE-alpha. It is not really functional, but the
main infrastructure is, to a large extent, in place. The main missing
piece is dealing with function (especially nested functions) definitions.
We will handle this by writing a very simple lambda lifter (this will
be dealt with as soon as possible).

Contents: the file names are from a previous project so they
are not very descriptive. I will rename them at some point in the
future.

  id.mli            symbol and unique name managment.
  id.ml
  parser.mly        ocamlyacc parser.
  lexer.mll         ocamllex lexer.
  parsetree.ml      ast.
  typedtree2.ml     this is the interemediate language produced
                    by the typechecker/translator Typecheck2
                    and consumed by the LLVM emitter EmitLlvm.
                    It will also be used by the Lambda lifter
                    once that is written. As a language it lives
                    essentially on the same level as LLVM, but it
                    has higher level control structures such as
                    conditions, loops, etc.
  typecheck2.ml     this is the main typechecker/translator.
  globals.ml        some useful global data structures.
  emitLlvm.ml       LLVM emitter.
  test/good         tests that the compiler should succeed at.
  test/bad          tests that the compiler should fail at.
  opt.mli           this handles the LLVM optimiser. It is not
  opt.ml            currently used.
  compile.ml        old driver routines
  tc.ml
  test.ml           current driver routine.
  myocamlbuild.ml   build system files
  test.mltop
  _tags
  Makefile
  .gitignore