Introduce ANF as intermediate language.
This module Anf will take the place of Typedtree2 later on.

The reasons to introduce a new intermediate language are
varied. The main one is aesthetic. The current intermediate
language is a very inelegant statement oriented language.
In particular the current intermediate language does not get
me access to all the features of the SSA language that underlies
LLVM (such as PHI-instructions). We could have used a language
that was closer to LLVM's, but then it is very hard to mantain
the scoping information for inner functions that we need to
perform Lambda Lifting (that we do after translation into the
intermediate language). The solutions is ANF, Administrative
Normal Form. It is the functional equivalent to SSA. We could have
also used CPS (Continuation Passing Style), but it is simpler to
generate LLVM code from ANF. The key feature is that it provides
for local blocks (like local function definitions, but that can
only be called in tail position). They correspond to basic blocks
in SSA. Its arguments correspond to PHI-nodes. All of this is
nicely explained in A. Appel's paper 'SSA is Functional Programming'.