Added Immutable flag to variables.
This is only used for iteration variables in FOR loops.
But it is crucial because otherwise the generated LLVM
code was not working (since iteration variables are immutables
we do not alloca them, just keep them in registers - so we must
remember not to LOAD them to get their values!).