Getting ready to switch to a one-pass generation of LLVM code.
We will do a separate typechecking phase and then generate
LLVM code *directly*. The main reason is simplicity. The current
code in Translate is very complicated. And while the generation
of the ANF intermediate representation is very beautiful, it is
not really useful for anything, since we do not do any type of
optimisation.

The module Typing will be in charge of typechecking and also of
computing which variables escape (are free variables of some
nested function) and of computing stack indices for these.

It already works on very simple example test/good/nested2.tig