Internal representation of the LLVM code.
Instead of emitting LLVM code directly using the bindings,
we will generate this internal representation of the LLVM code first.

This will be useful to have for debugging and also is more satisfactory
because we are no longer completely dependent on the bindings
(in the future we could also emit a text file if we wanted).

The only tricky part is that when we emit LLVM code there can be
cycles related to phi-nodes. These are not represented in this
representation, so their handling is a little delicate.