Add emitter of LLVM code for the representaiton IL.
One has to be a little careful to emit instructions
in the 'right' order so that all variables are defined.
This can be done by doing a depth-first search of the
basic block graph. The only tricly point is that there can
be cycles related to phi-nodes (if there are no phi-nodes
there are no problems, because undefined variables cannot
appear in the body of any basic block in well-formed LLVM
code).

Basically when we see a phi-node we emit all the incoming
values that are defined at that point. And we save the undefined
ones and we backpatch them later. The algorithm is written
in a purely functional way, hopefully it works, but it has
not yet been tested.