User-defined functions work again.
However, note that non-local variable access is not yet
implemented. In fact, there is a small problem w.r.t.
non-local variables which are gc roots:

a non-local variable will be stored in the stack inside
a structure. Some of the fields of this structure will be
gc roots, but is not clear that one can use a pointer to the
middle of such an structure as an argument to @llvm.gcroot.
For that reason I will go back to lambda lifting as a way to
handle non-local variables. I will implement it as a part of the
typechecking phase.