Compile FOR nodes without using PHI-instructions.
Since we do not use PHI-instructions anywhere else,
this makes the generated code more uniform. Moreover,
this might be useful for reintroducing one of the
intermediate languages we had used before (ANF or imperative IL)
because handling PHI-instructions in those was tricky
(as it can introduce cycles in the datastructures).